requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
numpy==1.22.3
torch==1.11.0
typing-extensions==4.1.1
pandas

# backwards compatibility
pathlib2
